@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.2
 
 environment:
-  sdk: '>=2.19.1 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -35,14 +35,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  # For google map and location
-  google_maps_flutter: ^2.2.6
-  google_maps_flutter_platform_interface: ^2.2.6
-  google_maps_flutter_android: ^2.4.6
-  permission_handler: ^10.2.0
+  google_maps_flutter: ^2.5.3
   geolocator: ^9.0.2
-  
-  #
   badges: ^3.1.1
   injectable: ^2.1.0
   json_annotation: ^4.8.0
@@ -51,19 +45,17 @@ dependencies:
   hive_flutter: ^1.1.0
   dio: ^5.1.2
   easy_localization: ^3.0.1
-  bloc: ^8.1.0
-  flutter_bloc: ^8.1.2
+  flutter_bloc: ^8.1.6
   async: ^2.9.0
   go_router: ^7.0.1
   equatable: ^2.0.5
   image_picker: ^0.8.7+5
   readmore: ^2.2.0
   flutter_svg: ^2.0.5
-  syncfusion_flutter_datepicker: 20.4.52
+  syncfusion_flutter_datepicker: 24.2.6
   image_cropper: ^3.0.1
   carousel_slider: ^4.2.1
   convert_vietnamese: ^0.0.9
-  
   path_provider: ^2.0.13
   http_parser: ^4.0.2
   mime: ^1.0.4
@@ -71,9 +63,12 @@ dependencies:
   path: ^1.8.2
   url_launcher: ^6.1.11
   shimmer: ^2.0.0
-  firebase_core: ^2.12.0
+  firebase_core: ^2.25.5
+  firebase_messaging: ^14.7.19
   firebase_analytics: ^10.4.1
   firebase_crashlytics: ^3.4.15
+  location: ^6.0.2
+  dash_chat_2: ^0.0.20
 
 dev_dependencies:
   flutter_test:
