@@ -26,9 +26,9 @@ export const CampaignResolver = Module.builder()
             route: '/organizations/:organizationId/campaigns',
             method: 'post',
             interceptors: [new MediaInterceptor(), CreateCampaignInterceptor],
-            body: 'CreateCampaignDto',
-            consumes: ['multipart/form-data'],
             params: [orgCampaignId, uploadMediaSwagger],
+            consumes: ['multipart/form-data'],
+            body: 'CreateCampaignDto',
             controller: CampaignController.createOne,
             preAuthorization: true,
         },
