@@ -1,5 +1,5 @@
 abstract class EnvKeys {
   static const String baseUrl = 'BASE_URL';
-  static const String googleMapApiKey = 'GOOGLE_MAP_API_KEY';
+  static const String googleMapApiKey = 'GOOGLE_MAPS_API_KEY';
   static const String localtionUrl = 'LOCATION_API_URL';
 }
