@@ -8,7 +8,7 @@ part of 'coordinate.model.dart';
 
 CoordinateModel _$CoordinateModelFromJson(Map<String, dynamic> json) =>
     CoordinateModel(
-      id: json['id'] as int,
+      id: (json['id'] as num).toInt(),
       name: json['name'] as String,
       coordinate: json['coordinate'] == null
           ? null
