@@ -1,6 +1,5 @@
 import 'dart:developer';
-
-import 'package:bloc/bloc.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:mobile/data/models/place.model.dart';
 import 'package:mobile/data/repositories/place.repository.dart';
