@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
@@ -9,7 +10,6 @@ import 'package:mobile/common/extensions/context.extension.dart';
 import 'package:mobile/common/utils/toast.util.dart';
 import 'package:mobile/common/widgets/custom_app_bar.widget.dart';
 import 'package:mobile/data/repositories/campaign.repository.dart';
-import 'package:mobile/data/repositories/place.repository.dart';
 import 'package:mobile/di/di.dart';
 import 'package:mobile/generated/locale_keys.g.dart';
 import 'package:mobile/modules/map/bloc/bottom_sheet_bloc/map_bottom_sheet.bloc.dart';
@@ -18,29 +18,21 @@ import 'package:mobile/modules/map/widgets/map_bottom_sheet.widget.dart';
 import 'package:mobile/modules/map/widgets/map_search_button.widget.dart';
 
 class MapPage extends StatelessWidget {
-  final Completer<GoogleMapController> controller = Completer();
+  final Completer<GoogleMapController> _controller =
+      Completer<GoogleMapController>();
 
   MapPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return MultiBlocProvider(
-      providers: [
-        BlocProvider(
-          create: (context) => MapBloc(
-            placeRepository: getIt.get<PlaceRepository>(),
-          ),
-        ),
-        BlocProvider(
-          create: (context) => MapBottomsheetBloc(
-            campaignRepository: getIt.get<CampaignRepository>(),
-          ),
-        ),
-      ],
+    return BlocProvider(
+      create: (context) => MapBottomsheetBloc(
+        campaignRepository: getIt.get<CampaignRepository>(),
+      ),
       child: BlocListener<MapBloc, MapState>(
         listener: (context, state) => _listenMapStateChanged(context, state),
         child: _MapView(
-          controller: controller,
+          controller: _controller,
         ),
       ),
     );
@@ -50,17 +42,23 @@ class MapPage extends StatelessWidget {
     BuildContext context,
     MapState state,
   ) async {
-    if (state is MapGetLocationSuccess) {
-      GoogleMapController googleMapController = await controller.future;
-      googleMapController.animateCamera(
-        CameraUpdate.newLatLngZoom(
-          state.myLocation ?? defaultLocation,
-          9,
-        ),
+    if (state.error != null && context.mounted) {
+      ToastUtil.showError(
+        context,
+        text: state.error,
       );
     }
-    if (state.error != null && context.mounted) {
-      ToastUtil.showError(context, text: state.error);
+
+    if (state is MapGetLocationSuccess) {
+      final controller = await _controller.future;
+      controller.animateCamera(
+        CameraUpdate.newCameraPosition(
+          CameraPosition(
+            target: state.myLocation ?? defaultLocation,
+            zoom: 9,
+          ),
+        ),
+      );
     }
   }
 }
@@ -93,7 +91,16 @@ class _MapView extends StatelessWidget {
               ),
               onMapCreated: (gController) {
                 controller.complete(gController);
+                context.read<MapBloc>().add(const MapMarkersGet());
               },
+              padding: Platform.isAndroid
+                  ? EdgeInsets.only(
+                      top: context.height * 0.2,
+                    )
+                  : EdgeInsets.only(
+                      bottom: context.height * 0.15,
+                    ),
+              myLocationEnabled: true,
               mapToolbarEnabled: false,
               zoomControlsEnabled: false,
               buildingsEnabled: false,
